extracted BackedEnum tests to separate file for PHP 7.4 compat
